Isolate Studio runtime configuration
